% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinCells.R
\name{BinCells}
\alias{BinCells}
\title{Bin scRNA-seq counts}
\usage{
BinCells(
  df,
  ann_df,
  group_col,
  join_col = "sample_id",
  n_bin = 10,
  bin_function = "median"
)
}
\arguments{
\item{df}{Data frame of normalized, filtered counts. The rows should be genes, and columns should be samples.}

\item{ann_df}{Data frame containing the group of each cell. One column should contain cell IDs, rest of the columns should contain corresponding groups of the cell}

\item{group_col}{Name of column in ann_df containing the group of each cell}

\item{join_col}{Name of column in ann_df containing the cell IDs}

\item{n_bin}{Number of random bins to partition in each group}

\item{bin_function}{The function to use when calculating the summary statistic}
}
\value{
Data frame with 1st column called "Geneid" containing gene names, rest of the columns containing the summary statistic for each bin. The bins are named by a random cell ID in the bin
}
\description{
Groups the cells, partitions each group into random bins, and calculates a summary statistic for the counts in each bin.
}
