% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinCells.R
\name{BootstrapSamples}
\alias{BootstrapSamples}
\title{Bootstrap scRNA-seq cells}
\usage{
BootstrapSamples(
  df,
  ann_df,
  group_col,
  join_col = "sample_id",
  n = 50,
  m = 100
)
}
\arguments{
\item{df}{Data frame of normalized, filtered counts. The rows should be genes, and columns should be samples.}

\item{ann_df}{Data frame containing the group of each cell. One column should contain cell IDs, rest of the columns should contain corresponding groups of the cell}

\item{group_col}{Name of column(s) in ann_df to use when grouping the cells. Can be a string (for one column) or vector (for multiple columns).}

\item{join_col}{Name of column in ann_df containing the cell IDs}

\item{n}{Number of cells to sample with replacement in each bootstrap sample}

\item{m}{Number of bootstrap samples to generate in each group}
}
\value{
A named list
\itemize{
\item samples - Data frame of bootstrapped samples, with first column as gene ID
\item ann - New ann_df, corresponding to the new sample IDs. Use this when ann_df is needed downstream.
}
}
\description{
Samples n cells with replacement from each group and taking the sum of their normalized counts
}
