% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preproc.R
\name{BulkPreproc}
\alias{BulkPreproc}
\title{Filter and normalize bulk RNA-seq data}
\usage{
BulkPreproc(df, cell_thresh = 2, count_thresh = 0.5)
}
\arguments{
\item{df}{Data frame with raw counts. The rows should be genes, and columns should be samples.}

\item{cell_thresh}{Minimum number of samples that need to have CPM above threshold (default=2)}

\item{count_thresh}{CPM threshold (default=0.5)}
}
\value{
data frame with normalized counts
}
\description{
Subsets the genes to those that have more than a threshold number of counts per million (CPM) in a threshold number of samples.
Then, normalizes the counts using the DESeq2 package.
}
