% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{convertSpecies}
\alias{convertSpecies}
\title{Convert gene IDs between species}
\usage{
convertSpecies(df, hm, old_id, new_id)
}
\arguments{
\item{df}{Data frame with raw counts. The rows should be genes, and columns should be samples}

\item{hm}{Homology table from Ensembl website, loaded as data frame}

\item{old_id}{Name of column in homology table containing the original species IDs}

\item{new_id}{Name of column in homology table containing the new species IDs}
}
\value{
New data frame with converted genes in the column called "Geneid"
}
\description{
Convert the IDs in the counts table to the IDs of the new species (one-to-one, high confidence homologs only)
}
