% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{convertSymbol}
\alias{convertSymbol}
\title{Gene symbol conversion}
\usage{
convertSymbol(
  df,
  gene_map,
  conv_col = "converted_alias",
  init_col = "initial_alias"
)
}
\arguments{
\item{df}{Data frame with raw counts. The rows should be genes, and columns should be samples.}

\item{gene_map}{Gene table from gProfiler, loaded as data frame}

\item{conv_col}{Name of column in gene_map with converted symbols/IDs}

\item{init_col}{Name of column in gene_map with original symbols/IDs}
}
\value{
New data frame with converted genes in the column called "Geneid"
}
\description{
Converts the gene symbols in the counts table to Ensembl IDs, or Ensembl IDs into gene symbols
}
