% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gqs_ann}
\alias{gqs_ann}
\title{Growing, quiescent, senescent dataset annotations}
\format{
A data frame with 8 rows and 2 columns:
\describe{
\item{sample_id}{IDs of the cells}
\item{Condition}{The state of the samples (growing, quiescence, senescence, deep senescence)}
}
}
\usage{
gqs_ann
}
\description{
The annotations for the growing, quiescent, senescent bulk RNA-seq dataset
}
\keyword{datasets}
