% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gqs_raw}
\alias{gqs_raw}
\title{Growing, quiescent, senescent example dataset}
\format{
A data frame with 62703 rows and 9 columns:
\describe{
\item{Geneid}{Ensembl IDs of genes}
\item{SRR9601022}{Raw counts for 1st sample}
...
}
}
\usage{
gqs_raw
}
\description{
A dataset containing the raw counts for bulk RNA-seq of fibroblasts
under growing, quiescent, senescent, and deep senescent conditions, with 2 samples
per condition
}
\keyword{datasets}
