% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDSFunctions.R
\name{make_boxplot}
\alias{make_boxplot}
\title{Make a boxplot of quiescence depth predictions}
\usage{
make_boxplot(
  df,
  ann_df,
  x,
  y = "QDS",
  byvar = c(sample_id = "sample_id"),
  color = NULL,
  title = NULL,
  levels = NULL
)
}
\arguments{
\item{df}{Data frame with quiescence depth of each sample. One column should contain the sample IDs, and the other should contain the predicted values}

\item{ann_df}{Data frame with one column containing sample IDs, rest of the columns containing the corresponding group of the sample}

\item{x}{Name of column in ann_df with the group of the samples}

\item{y}{Name of column in df containing the predicted quiescence depth values}

\item{byvar}{Columns to use when joining df and ann_df. Example: c("df_column"="ann_column")}

\item{color}{Name of column to use when coloring the points on the boxplot}

\item{title}{Title of plot}

\item{levels}{Vector containing the order of groups on the plot}
}
\value{
Grouped boxplot of quiescence depth of test samples
}
\description{
Groups the test samples and makes a boxplot of the quiescence depth values in each group.
There is the option to color the points on the boxplot
}
