% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDSFunctions.R
\name{make_grouped_boxplot}
\alias{make_grouped_boxplot}
\title{Make a grouped boxplot of quiescence depth predictions}
\usage{
make_grouped_boxplot(
  df,
  ann_df,
  x1,
  x2,
  y = "QDS",
  byvar = c(sample_id = "sample_id"),
  title = NULL,
  levels = NULL
)
}
\arguments{
\item{df}{Data frame with quiescence depth of each sample. One column should contain the sample IDs, and the other should contain the predicted values}

\item{ann_df}{Data frame with one column containing sample IDs, rest of the columns containing the corresponding group of the sample}

\item{x1}{Name of column in ann_df with the first condition to use when grouping the samples}

\item{x2}{Name of column in ann_df with the second condition to use when grouping the samples}

\item{y}{Name of column in df containing the predicted quiescence depth values}

\item{byvar}{Columns to use when joining df and ann_df. Example: c("df_column"="ann_column")}

\item{title}{Title of plot}

\item{levels}{Vector containing the order of groups (x1) on the plot}
}
\value{
Grouped boxplot of quiescence depth of test samples
}
\description{
Groups the test samples by 2 conditions and makes a boxplot of the quiescence depth values in each condition.
}
