% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDSFunctions.R
\name{make_grouped_hist}
\alias{make_grouped_hist}
\title{Plot density curves of scRNA-seq quiescence depth predictions}
\usage{
make_grouped_hist(
  pred_df,
  ann_df,
  grouping1,
  xvar = "QDS",
  grouping2 = NULL,
  byvar = c(sample_id = "sample_id"),
  title = NULL,
  levels = NULL
)
}
\arguments{
\item{pred_df}{Data frame with quiescence depth of each sample. One column should contain the sample IDs, and the other should contain the predicted values}

\item{ann_df}{Data frame with one column containing sample IDs, rest of the columns containing the corresponding group of the sample}

\item{grouping1}{The name of the column in ann_df with the group of each sample}

\item{xvar}{the column in pred_df containing the predicted values}

\item{grouping2}{the name of the column in ann_df to use when splitting the density curves in each group}

\item{byvar}{Columns to use when joining df and ann_df. Example: c("df_column"="ann_column")}

\item{title}{Title of plot}

\item{levels}{Vector containing the order of groups on the plot}
}
\value{
Density curves that show the distribution of quiescence depth in groups of single cells
}
\description{
Groups the test samples and plots density curves for each group. There is the option to plot multiple density curves for each group
}
