% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeHist.R
\name{MakeHist}
\alias{MakeHist}
\title{Make histograms of QC metric distributions}
\usage{
MakeHist(df, ann_df = NULL, mt.pattern = "^MT-", grouping = NULL)
}
\arguments{
\item{df}{Data frame with raw counts. The rows should be genes, and columns should be samples}

\item{ann_df}{Metadata table containing biological condition of each cell (optional)}

\item{mt.pattern}{regex pattern for mitochondrial genes (default: "^MT-")}

\item{grouping}{Name of variable to group the histograms by (optional)}
}
\value{
List of histograms (ggplot2 objects) for nUMI, nGene, log10GenesPerUMI, and mitoRatio
}
\description{
The Make_Hist() function calculates total raw counts (nUMI), total number of non zero genes (nGene),
non genes per total raw counts (log10GenesPerUMI), and mitochondrial gene counts per total raw counts (mitoRatio)
and plots a histogram of each.
}
