% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_rat_hm}
\alias{mouse_rat_hm}
\title{Homology table for mouse and rat Ensembl IDs}
\format{
A data frame with 128614 rows and 2 columns:
\describe{
\item{Gene stable ID}{Mouse Ensembl gene IDs}
\item{Gene stable ID version}{Mouse Ensembl gene IDs, with version}
\item{Transcript stable ID}{Mouse Ensembl transcript IDs}
\item{Transcript stable ID version}{Mouse Ensembl transcript IDs, with version}
\item{Rat gene stable ID}{Rat Ensembl gene IDs}
\item{Rat orthology confidence}{Confidence in homology relationship (0=low, 1=high)}
\item{Rat homology type}{Homology type (one-to-one, one-to-many, many-to-many)}
}
}
\usage{
mouse_rat_hm
}
\description{
The corresponding Ensembl IDs for mouse and rat, as well as the homology type and confidence
}
\keyword{datasets}
