% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsc_ann}
\alias{nsc_ann}
\title{Neural stem cell (NSC) dataset annotations}
\format{
A data frame with 2785 rows and 3 columns:
\describe{
\item{sample_id}{IDs of the cells}
\item{celltype}{The dormancy state (dormant, primed, activated (early), or activted (late) )}
\item{age}{The age (old or young)}
}
}
\usage{
nsc_ann
}
\description{
The annotations for the NSC scRNA-seq dataset
}
\keyword{datasets}
