% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsc_genemap}
\alias{nsc_genemap}
\title{Gene map for neural stem cell (NSC) dataset}
\format{
A data frame with 27998 rows and 2 columns:
\describe{
\item{initial_alias}{gene names}
\item{converted_alias}{Ensembl IDs}
}
}
\usage{
nsc_genemap
}
\description{
The Ensembl IDs corresponding to the gene names in the NSC dataset
}
\keyword{datasets}
