% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsc_raw}
\alias{nsc_raw}
\title{Neural stem cell (NSC) example dataset}
\format{
A data frame with 27998 rows and 2785 columns:
\describe{
\item{AAACCTGCACCAGTTA-1}{Raw counts for 1st sample}
\item{AAACCTGCATGGTCAT-1}{Raw counts for 2nd sample}
...
}
}
\usage{
nsc_raw
}
\description{
A dataset containing the raw counts for scRNA-seq of
dormant, primed, early-stage activated, and late-stage activated neural stem cells (NSC),
with 2785 total cells
}
\keyword{datasets}
