% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pd_ann}
\alias{pd_ann}
\title{16-74 population doublings dataset annotations}
\format{
A data frame with 15 rows and 2 columns:
\describe{
\item{sample_id}{IDs of the cells}
\item{population_doublings}{The number of population doublings the sample has gone through}
}
}
\usage{
pd_ann
}
\description{
The annotations for the 16-74 population doublings bulk RNA-seq dataset
}
\keyword{datasets}
