% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pd_raw}
\alias{pd_raw}
\title{Population doubling example dataset}
\format{
A data frame with 61541 rows and 16 columns:
\describe{
\item{Geneid}{Ensembl IDs of genes}
\item{SRR1660543}{Raw counts for 1st sample}
...
}
}
\usage{
pd_raw
}
\description{
A dataset containing the raw counts for bulk RNA-seq of fibroblasts
at 16, 24, 46, 64, and 72 population doublings, with 3 samples per condition
}
\keyword{datasets}
