% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDS.R
\name{QDS}
\alias{QDS}
\title{Train a model and predict quiescence depth}
\usage{
QDS(
  test_df,
  train_df = NULL,
  ann_df = NULL,
  y_col = NULL,
  prebuilt = F,
  return.all = F
)
}
\arguments{
\item{test_df}{test data (normalized counts). The rows should be genes, and columns should be samples.}

\item{train_df}{train data (normalized counts). The rows should be genes, and columns should be samples.}

\item{ann_df}{data frame with train data labels. First column contains train sample IDs and 2nd column contains corresponding y values.}

\item{y_col}{Name of column in ann_df that contains y values}

\item{prebuilt}{Set to TRUE if using the prebuilt rat embryonic fibroblast model to predict on the test data.  If prebuilt=TRUE, provide only the test_df argument. Leave train_df, ann_df, and y_col arguments empty.}
}
\value{
Data frame with predicted quiescence depth value for each test sample
}
\description{
Merges the train and test datasets and removes batch effects between them.
Then, builds an elastic net regression model on the train data. In elastic net, the alpha parameter controls how many coefficients (genes) are in the model:
lower alpha gives models with more coefficients, higher alpha gives models with fewer coefficients.
Tests a range of alpha values and finds the cross-validation error of each one. T
To capture a broader signature of quiescence depth, we use the minimum alpha with cross-validation error within 5\% of the minimum error.
}
