% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rat_human_hm}
\alias{rat_human_hm}
\title{Homology table for human and rat Ensembl IDs}
\format{
A data frame with 179981 rows and 2 columns:
\describe{
\item{Gene stable ID}{Human Ensembl gene IDs}
\item{Gene stable ID version}{Human Ensembl gene IDs, with version}
\item{Transcript stable ID}{Human Ensembl transcript IDs}
\item{Transcript stable ID version}{Human Ensembl transcript IDs, with version}
\item{Rat gene stable ID}{Rat Ensembl gene IDs}
\item{Rat orthology confidence}{Confidence in homology relationship (0=low, 1=high)}
\item{Rat homology type}{Homology type (one-to-one, one-to-many, many-to-many)}
}
}
\usage{
rat_human_hm
}
\description{
The corresponding Ensembl IDs for human and rat, as well as the homology type and confidence
}
\keyword{datasets}
