% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_ann}
\alias{ref_ann}
\title{Rat embryonic fibroblast (REF) dataset annotations}
\format{
A data frame with 27 rows and 3 columns:
\describe{
\item{sample_id}{IDs of the cells}
\item{Condition}{The dormancy state (all Q, for quiescent))}
\item{Dataset}{The dataset (all REF)}
\item{Day}{The number of serum starvation days (2-16)}
}
}
\usage{
ref_ann
}
\description{
The annotations for the bulk RNA-seq of
rat embryonic fibroblasts (REF) under 2, 3, 4, 6, 8, 10, 12, 14, and 16
days of serum starvation
}
\keyword{datasets}
