% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_raw}
\alias{ref_raw}
\title{Rat embryonic fibroblast (REF) example dataset}
\format{
A data frame with 30560 rows and 28 columns:
\describe{
\item{Geneid}{Ensembl IDs of genes}
\item{SRR8353399}{Raw counts for 1st sample}
...
}
}
\usage{
ref_raw
}
\description{
A dataset containing the raw counts for bulk RNA-seq of
rat embryonic fibroblasts (REF) under 2, 3, 4, 6, 8, 10, 12, 14, and 16
days of serum starvation, with 3 samples per group
}
\keyword{datasets}
