% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preproc.R
\name{scPreproc}
\alias{scPreproc}
\title{Filter and normalize scRNA-seq data}
\usage{
scPreproc(
  df,
  prop = 0.002,
  nUMI_filt = 500,
  nGene_filt = 250,
  log10_filt = 0.75,
  mito_filt = 0.2,
  variable.features = 10000
)
}
\arguments{
\item{df}{Data frame with raw counts. The rows should be genes, and columns should be samples.}

\item{prop}{Threshold proportion of cells with non zero counts}

\item{nUMI_filt}{nUMI (total raw counts (nUMI) threshold}

\item{nGene_filt}{nGene (total number of non zero genes) threshold}

\item{log10_filt}{log10GenesPerUMI (non zero genes per total raw counts) threshold}

\item{mito_filt}{mitoRatio (mitochondrial gene counts per total raw counts) threshold}

\item{variable.features}{SCTransform subsets the dataset to the top n variable genes. This argument specifies n}
}
\value{
data frame with normalized counts
}
\description{
Filters the cells by the nUMI, nGene, log10GenesPerUMI, and mitoRatio threshold values,
gets the genes with non zero counts in X or more cells, where X=total cells * threshold proportion,
uses SCTransform to normalize the raw counts. Uses vst.flavor="v2",
and subsets the normalized counts table to the genes that pass threshold
}
