% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preproc.R
\name{FilterDoublets}
\alias{FilterDoublets}
\title{Filter doublets from scRNA-seq data}
\usage{
FilterDoublets(df, ann_df, split.by, byvar)
}
\arguments{
\item{df}{Data frame with raw counts. The rows should be genes, and columns should be samples.}

\item{ann_df}{Metadata table containing the sample each cell belongs to. The cells are grouped by sample, and the doublet-finding algorithm is run on each group separately}

\item{split.by}{Name of column(s) in ann_df to use when grouping the cells. Can be a string (for one column) or vector (for multiple columns).}

\item{id_col}{Name of column in ann_df containing the cell IDs}
}
\value{
Raw counts table with doublets filtered
}
\description{
This function uses the DoubletFinder package to find doublets
and removes the doublets from the data table
}
